steganocrypt/requirements.txt
Pillow==10.3.0
pycryptodomex==3.23.0
mnemonic